app/tracking_project/requirements.txt
attrs==25.3.0
blinker==1.9.0
click==8.3.0
flasgger==0.9.7.1
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.2
mistune==3.1.4
packaging==25.0
PyMySQL==1.1.2
python-dotenv==1.1.1
PyYAML==6.0.3
referencing==0.36.2
rpds-py==0.27.1
six==1.17.0
SQLAlchemy==2.0.43
typing_extensions==4.15.0
Werkzeug==3.1.3
